+++ src/adapter/coinmarketcap/new_listing.rs
@@ -1,0 +1,8 @@
+use super::prelude::*;
+
+pub async fn run_signal_new_listing() -> Result<()> {
+    // historic listing
+    todo!("implement new listing");
+    // let url = "https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest";
+    // Ok(())
+}