+++ src/lib.rs
@@ -2,9 +2,11 @@
 pub mod adapter;
 // api server
 pub mod api;
+// project model
+pub mod model;
 // toml configuration
 pub mod config;
-// project custom error
+// project error
 pub mod error;
 // metadata
 pub mod metadata;