+++ src/main.rs
@@ -1,6 +1,6 @@
 use futures::future::join_all;
 use futures::TryFutureExt;
-use reqwest;
+use signals::adapter::coinmarketcap::run_signal_fear_and_index;
 use signals::api::host_rest_api_server;
 use signals::config::Config;
 use signals::error::{Result, SignalsError};
@@ -27,7 +27,8 @@ async fn main() -> Result<()> {
     if config.has_runner("coinmarketcap") {
         println!("Starting CoinMarketCap indexer");
         let period = tokio::time::Duration::from_secs(10);
-        let handle = tokio::spawn(async move { run_coinmarket_cap(config.clone(), period).await });
+        let handle =
+            tokio::spawn(async move { run_signal_fear_and_index(config.clone(), period).await });
         runners.push(handle);
     }
 
@@ -37,37 +38,4 @@ async fn main() -> Result<()> {
         };
     }
     Ok(())
-}
-
-// TODO set up the global source to replace the config
-async fn run_coinmarket_cap(config: Config, period: tokio::time::Duration) -> Result<()> {
-    let mut interval = tokio::time::interval(period);
-
-    // set up coinmarket cap fetchers
-    let client = reqwest::Client::new();
-
-    // historic listing
-    let url = "https://pro-api.coinmarketcap.com/v1/cryptocurrency/listings/latest";
-    // fear and greed
-    let url = "https://pro-api.coinmarketcap.com/v3/fear-and-greed/historical";
-    // bitcoin dominance
-    let url = "https://pro-api.coinmarketcap.com/v1/global-metrics/quotes/latest";
-
-    let key_cmc = config.get_api_key("coinmarketcap")?;
-
-    loop {
-        interval.tick().await; // Wait for the next tick
-
-        println!("Fetching coinmarket cap:");
-        // dispatch request
-
-        let response = client
-            .get(url)
-            .header("X-CMC_PRO_API_KEY", &key_cmc)
-            .send()
-            .await
-            .unwrap();
-        let text = response.text().await.unwrap();
-        println!("{:?}", text);
-    }
 }