+++ src/model/mod.rs
@@ -1,0 +1,7 @@
+// signal -> strategy -> instance
+/// instance
+pub mod instance;
+/// signal
+pub mod signal;
+/// strategy
+pub mod strategy;