+++ src/model/strategy.rs
@@ -1,0 +1,6 @@
+pub struct StrategyInfo {
+    id: u64,
+    name: String,
+    description: String,
+    creator: String,
+}